pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lineage"
version = "0.1.0"
description = "Git blame for AI-generated code - EU AI Act compliance tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Syed Arfan", email = "arfan@example.com"}
]
keywords = ["ai", "compliance", "eu-ai-act", "git", "attribution", "audit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "flask>=3.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
lineage = "lineage.cli.main:cli"

[project.urls]
Homepage = "https://github.com/simplyarfan/Lineage"
Documentation = "https://github.com/simplyarfan/Lineage/blob/main/docs/ARCHITECTURE.md"
Repository = "https://github.com/simplyarfan/Lineage"
Issues = "https://github.com/simplyarfan/Lineage/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["lineage*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]